functions from fast1.cpp moved into fst.cpp
